Update Installation to better support

+++ install/install1.php
@@ -1,4 +1,6 @@
 <?php
+define( 'APPLICATION_LOADED', true );
+
 /*********************************************************************
 PHPBack
 Ivan Diaz <ivan@phpback.org>
@@ -7,12 +9,13 @@ http://www.phpback.org
 Released under the GNU General Public License WITHOUT ANY WARRANTY.
 See LICENSE.TXT for details.
 **********************************************************************/
+define('BASEPATH', '');
 
 error_reporting(E_ALL);
 ini_set('display_errors', 1);
 
 function redirectpost($url, array $data){
-    	
+
 		echo '<html>
 		    <head>
 		        <script type="text/javascript">
@@ -42,12 +45,10 @@ function exit_error($string){
 }
 
 function create_file($hostname, $username, $password, $database){
-	chmod('../application/config', 0777);
-	if(($file = fopen('../application/config/database.php', 'w')) == FALSE){
-		$data['error'] == "Could not create the config file";
-		$data['configfile'] == 1;
-		redirectpost('index.php', $data);
-	}
+	@chmod('../application/config', 0777);
+	if(($file = fopen('../application/config/database.php', 'w+')) == FALSE){
+        exit_error('ERROR #1: Config file could not be created');
+    }
 	$content = '<?php ';
 	$content .= '$active_group = \'default\';';
 	$content .= '$active_record = TRUE;';
@@ -67,11 +68,11 @@ function create_file($hostname, $username, $password, $database){
 	$content .= '$db[\'default\'][\'autoinit\'] = TRUE;';
 	$content .= '$db[\'default\'][\'stricton\'] = FALSE;';
 
-	if(fwrite($file, $content) == FALSE){
-		$data['error'] == "Could not create the config file";
-		$data['configfile'] == 1;
-		redirectpost('index.php', $data);
-	}
+	if(fwrite($file, $content) == FALSE) {
+        fclose($file);
+        exit_error('ERROR #1: Config file could not be created');
+    }
+
 	fclose($file);
 }
 
@@ -80,7 +81,7 @@ if($_POST['adminpass'] != $_POST['adminrpass'])
 
 $server = new mysqli($_POST['hostname'], $_POST['username'], $_POST['password']);
 
-if ($server->connect_error) exit_error('Connection Error (' . $server->connect_errno . ') ' . $server->connect_error);
+if ($server->connect_error) exit_error('ERROR #2: Server connection error (' . $server->connect_errno . ') ' . $server->connect_error);
 
 if($_POST['database'] != ""){
 	if(!file_exists('../application/config/database.php'))
@@ -88,33 +89,52 @@ if($_POST['database'] != ""){
 	include '../application/config/database.php';
 	if(!($_POST['hostname'] == $db['default']['hostname'] && $_POST['username'] == $db['default']['username']
 		&& $_POST['password'] == $db['default']['password'] && $_POST['database'] == $db['default']['database'])) exit_error('Config file does not match with the given information');
-	if ($server->select_db($_POST['database']) === FALSE) exit_error("Couldn't connect to database");
+	if ($server->select_db($_POST['database']) === FALSE) exit_error("ERROR #3: Couldn't connect to database");
 	$query = file_get_contents('database_tables.sql');
-	if($server->multi_query($query) === FALSE) exit_error("Couldn't create the tables");
+	if($server->multi_query($query) === FALSE) exit_error("ERROR #4: Couldn't create the tables");
 }else{
 	if(!file_exists('../application/config/database.php'))
 		create_file($_POST['hostname'], $_POST['username'], $_POST['password'], 'phpback');
-	if(!$server->query("CREATE DATABASE IF NOT EXISTS phpback")){
-		exit_error("Could not create database");
+
+    if ($server->select_db('phpback') === TRUE) exit_error("ERROR #5: You already have a phpback database, please create another manually");
+    if(!$server->query("CREATE DATABASE IF NOT EXISTS phpback")){
+		exit_error("ERROR #6: Could not create database");
 	}
-	if ($server->select_db('phpback') === FALSE) exit_error("Couldn't connect to database");
+	if ($server->select_db('phpback') === FALSE) exit_error("ERROR #5: Generated database connection error");
 	$sql = file_get_contents('database_tables.sql');
-	if($server->multi_query($sql) === FALSE) exit_error("Couldn't create the tables");
+	if($server->multi_query($sql) === FALSE) exit_error("ERROR #4: Couldn't create the tables");
 }
 do{
 	if($r = $server->store_result()) $r->free();
-}while($server->next_result());
+}while($server->more_results() && $server->next_result());
 
-unlink('index.php');
-unlink('install1.php');
-unlink('database_tables.sql');
 $result = $server->query("SELECT id FROM settings WHERE name='title'");
-if($result->num_rows == 1){
-	unlink('index2.php');
-	unlink('install2.php');
-	header('Location: ../admin');
-	exit;
+
+if ($result->num_rows == 1) {
+    if (!@chmod('../install', 0777)) {
+        echo "PLEASE DELETE install/ FOLDER MANUALLY. THEN GO TO yourwebsite.com/feedback/admin/ TO LOG IN.";
+        exit;
+    }
+
+    unlink('index.php');
+    unlink('install1.php');
+    unlink('database_tables.sql');
+    unlink('index2.php');
+    unlink('install2.php');
+    header('Location: ../admin');
+    exit;
+
+} else {
+    $server->query("INSERT INTO users(id,name,email,pass,votes,isadmin,banned) VALUES('','". $_POST['adminname'] ."','". $_POST['adminemail'] ."','". crypt($_POST['adminpass']) ."', 20, 3,0)");
+
+    if (!@chmod('../install', 0777)) {
+        echo "PLEASE DELETE install/index.php, install/install1.php AND install/database_tables.sql FILES MANUALLY.<br />
+            THEN GO TO yourwebsite.com/feedback/install/index2.php TO CONTINUE THE INSTALLATION.";
+        exit;
+    }
+
+    unlink('index.php');
+    unlink('install1.php');
+    unlink('database_tables.sql');
+    header('Location: index2.php');
 }
-$server->query("INSERT INTO users(id,name,email,pass,votes,isadmin,banned) VALUES('','". $_POST['adminname'] ."','". $_POST['adminemail'] ."','". crypt($_POST['adminpass']) ."', 20, 3,0)");
-header('Location: index2.php');
-exit;
